rag-standard-master/requirements.txt
ipykernel
ipywidgets
nbformat
setproctitle
python-dotenv
PyYAML
dill
pandas
pyarrow
fastparquet
matplotlib
tabulate
openpyxl
xlrd
python-docx
python-pptx
olefile
lxml
pdfminer.six
pdfplumber
PyMuPDF
pymupdf4llm
Markdown
markdownify
htmlmin
torch
transformers
nltk
kiwipiepy
rouge-score
openai
langchain
langgraph
fastembed
qdrant-client
opensearch-py
fastapi
uvicorn
beautifulsoup4